Contact Info & Map Widget: Remove Geocoding and add map to form
The Google Maps Geocoding API is used to verify that the given address
is valid. However, use of this API has recently caused problems for
users. Users need to enable two Google Maps APIs in order to use this
widget, which isn't user-friendly.

Instead of using the Geocoding API results to verify the address, show
the embedded map in the widget's admin form. The user can verify that the
address is valid and correct by looking at the map. Also, the user will
receive Google Maps' error messages if the API key is not valid.

Do not show the embedded map In the Customizer admin form. The embedded
map isn't necessary in this form because the embedded map is shown in
the Customizer preview.